go/roothash: Move RuntimeState to public API
